+++ src/Tecan/Rack/ScannedRack.php
@@ -1,0 +1,5 @@
+<?php declare(strict_types=1);
+
+namespace MLL\Utils\Tecan\Rack;
+
+interface ScannedRack extends Rack {}