+++ tests/Tecan/BasicCommands/AspirateTest.php
@@ -7,6 +7,7 @@ use MLL\Utils\Tecan\LiquidClass\CustomLiquidClass;
 use MLL\Utils\Tecan\LiquidClass\MLLLiquidClass;
 use MLL\Utils\Tecan\Location\BarcodeLocation;
 use MLL\Utils\Tecan\Location\PositionLocation;
+use MLL\Utils\Tecan\Rack\FluidXRack;
 use MLL\Utils\Tecan\Rack\MLLLabWareRack;
 use PHPUnit\Framework\TestCase;
 
@@ -15,10 +16,10 @@ final class AspirateTest extends TestCase
     public function testAspirateWithBarcodeLocation(): void
     {
         $barcode = 'barcode';
-        $aspirate = new Aspirate(100, new BarcodeLocation($barcode, MLLLabWareRack::DEST_PCR()), new CustomLiquidClass('TestLiquidClassName'));
+        $aspirate = new Aspirate(100, new BarcodeLocation($barcode, new FluidXRack()), new CustomLiquidClass('TestLiquidClassName'));
         self::assertSame($barcode, $aspirate->location->tubeID());
         self::assertNull($aspirate->location->position());
-        self::assertSame('A;;;96 Well PCR ABI semi-skirted;;barcode;100;TestLiquidClassName;;', $aspirate->toString());
+        self::assertSame('A;;;96FluidX;;barcode;100;TestLiquidClassName;;', $aspirate->toString());
     }
 
     public function testAspirateWithPositionLocation(): void