+++ tests/Tecan/TecanProtocolTest.php
@@ -17,6 +17,7 @@ use MLL\Utils\Tecan\CustomCommands\TransferWithAutoWash;
 use MLL\Utils\Tecan\LiquidClass\CustomLiquidClass;
 use MLL\Utils\Tecan\LiquidClass\MLLLiquidClass;
 use MLL\Utils\Tecan\Location\BarcodeLocation;
+use MLL\Utils\Tecan\Rack\FluidXRack;
 use MLL\Utils\Tecan\Rack\MLLLabWareRack;
 use MLL\Utils\Tecan\TecanProtocol;
 use MLL\Utils\Tecan\TipMask\TipMask;
@@ -71,7 +72,7 @@ CSV
         $tecanProtocol = new TecanProtocol(TipMask::FOUR_TIPS());
 
         $liquidClass = new CustomLiquidClass('TestLiquidClassName');
-        $rack = MLLLabWareRack::DEST_LC();
+        $rack = new FluidXRack();
         $aspirateLocation = new BarcodeLocation('barcode', $rack);
         $dispenseLocation = new BarcodeLocation('barcode1', $rack);
 
@@ -83,21 +84,21 @@ CSV
         self::assertSame(
             StringUtil::normalizeLineEndings(
                 <<<CSV
-{$this->initComment()}A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;1
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;1
+{$this->initComment()}A;;;96FluidX;;barcode;100;TestLiquidClassName;;1
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;1
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;2
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;2
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;2
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;2
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;4
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;4
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;4
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;4
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;8
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;8
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;8
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;8
 W;
 B;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;1
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;1
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;1
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;1
 W;
 
 CSV
@@ -111,7 +112,7 @@ CSV
         $tecanProtocol = new TecanProtocol(TipMask::FOUR_TIPS());
 
         $liquidClass = new CustomLiquidClass('TestLiquidClassName');
-        $rack = MLLLabWareRack::DEST_LC();
+        $rack = new FluidXRack();
         $aspirateLocation = new BarcodeLocation('barcode', $rack);
         $dispenseLocation = new BarcodeLocation('barcode1', $rack);
 
@@ -130,21 +131,21 @@ CSV
         self::assertSame(
             StringUtil::normalizeLineEndings(
                 <<<CSV
-{$this->initComment()}A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;1
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;1
+{$this->initComment()}A;;;96FluidX;;barcode;100;TestLiquidClassName;;1
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;1
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;2
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;2
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;2
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;2
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;4
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;4
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;4
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;4
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;8
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;8
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;8
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;8
 W;
 B;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;1
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;1
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;1
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;1
 W;
 
 CSV
@@ -158,7 +159,7 @@ CSV
         $tecanProtocol = new TecanProtocol(TipMask::EIGHT_TIPS());
 
         $liquidClass = new CustomLiquidClass('TestLiquidClassName');
-        $rack = MLLLabWareRack::DEST_LC();
+        $rack = new FluidXRack();
         $aspirateLocation = new BarcodeLocation('barcode', $rack);
         $dispenseLocation = new BarcodeLocation('barcode1', $rack);
 
@@ -171,36 +172,36 @@ CSV
         self::assertSame(
             StringUtil::normalizeLineEndings(
                 <<<CSV
-{$this->initComment()}A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;1
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;1
+{$this->initComment()}A;;;96FluidX;;barcode;100;TestLiquidClassName;;1
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;1
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;2
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;2
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;2
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;2
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;4
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;4
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;4
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;4
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;8
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;8
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;8
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;8
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;16
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;16
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;16
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;16
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;32
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;32
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;32
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;32
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;64
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;64
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;64
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;64
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;128
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;128
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;128
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;128
 W;
 B;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;1
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;1
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;1
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;1
 W;
-A;;;96 Well MP LightCycler480;;barcode;100;TestLiquidClassName;;2
-D;;;96 Well MP LightCycler480;;barcode1;100;TestLiquidClassName;;2
+A;;;96FluidX;;barcode;100;TestLiquidClassName;;2
+D;;;96FluidX;;barcode1;100;TestLiquidClassName;;2
 W;
 
 CSV